refactor polls
